pyproject.toml
[tool.poetry]
name = "django-bookstore-website"
version = "0.1.0"
description = ""
authors = ["Takib Yeasar <mdtakibyeasar@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.9"
pillow = "^10.4.0"
setuptools = "^70.3.0"
python-dotenv = "^1.0.1"
coverage = "^7.5.4"
flake8 = "^7.1.0"
django-countries = "^7.6.1"
stripe = "^10.3.0"
black = "^24.4.2"
isort = "^5.13.2"
django-debug-toolbar = "^4.4.6"
django-mptt = "^0.16.0"
paypal-checkout-serversdk = "^1.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
